removed unnecessary method bindings and reorganized utils
