+++ utils/request.js
@@ -6,7 +6,7 @@ const fetch = require('fetch');
  *
  * @return {object}          The parsed JSON from the request
  */
-function parseJSON (response) {
+const parseJSON = (response) => {
   if (response.status === 204 || response.status === 205) {
     return null;
   }
@@ -20,7 +20,7 @@ function parseJSON (response) {
  *
  * @return {object|undefined} Returns either the response, or throws an error
  */
-function checkStatus (response) {
+const checkStatus = (response) => {
   if (response.status >= 200 && response.status < 300) {
     return response;
   }
@@ -38,8 +38,8 @@ function checkStatus (response) {
  *
  * @return {object}           The response data
  */
-module.exports = function request (url, options) {
+module.exports = (url, options) => {
   return fetch.fetchUrl(url, options)
     .then(checkStatus)
     .then(parseJSON);
-}
+};
